build: fails installing et_xmlfile
The latest version of pip resolves to a version of et_xmlfile that fails to
install.  The workaround is to pin pip to 20.1.1.  Debian is upgraded from
version 9 to 10.  This results in python 3.7 being used instead of 3.5, which
has been 'end of life.'  npm dependencies have been updated for critical
fixes.

Ref : #87